% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXTOTU}
\alias{ggiNEXTOTU}
\title{ggplot2 extension for an iNEXT.OTU object}
\usage{
ggiNEXTOTU(output, type = "B", scale = "fixed", transp = 0.4)
}
\arguments{
\item{output}{the output from iNEXTOTU}

\item{type}{(required only when \code{base = "coverage"}), selection of plot type : \cr
\code{type = 'B'} for plotting the gamma, alpha, and beta diversity ;  \cr
\code{type = 'D'} for plotting 4 turnover dissimilarities.}

\item{scale}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{transp}{a value between 0 and 1 for controlling transparency. \code{transp = 0} is completely transparent, default is 0.4.}
}
\value{
a figure for phylogenetic diversity decomposition or dissimilarity measure.
}
\description{
\code{ggiNEXTOTU}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXTOTU}} object to plot coverage- or sample-sized-based rarefaction/extrapolation curves for phylogenetic diversity decomposition and dissimilarity measure.
}
\examples{

data("tongue_cheek")
data("tongue_cheek_tree")
output <- iNEXTOTU(tongue_cheek, q=c(0,1,2), nboot = 0, PDtree = tongue_cheek_tree)
ggiNEXTOTU(output, scale = 'free', transp = 0.4)

}
