% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXTUniFrac}
\alias{ggiNEXTUniFrac}
\title{ggplot2 extension for an iNEXT.UniFrac object}
\usage{
ggiNEXTUniFrac(output, scale = "fixed", transp = 0.4)
}
\arguments{
\item{output}{the output from iNEXTUniFrac}

\item{scale}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{transp}{a value between 0 and 1 for controlling transparency. \code{transp = 0} is completely transparent, default is 0.4.}
}
\value{
a figure for two types of UniFrac based on dissimilarity measure.
}
\description{
\code{ggiNEXTUniFrac}: the \code{\link[ggplot2]{ggplot}} extension for \code{\link{iNEXTUniFrac}} object to plot coverage-based rarefaction/extrapolation curves for UniFrac.
}
\examples{

data("tongue_cheek")
data("tongue_cheek_tree")
output <- iNEXTUniFrac(tongue_cheek, q=c(0,1,2), nboot = 0, PDtree = tongue_cheek_tree)
ggiNEXTUniFrac(output, scale = 'free', transp = 0.4)

}
